agrego excepciones

+++ src/main/java/com/aninfo/proyectos/service/ProyectoService.java
@@ -1,5 +1,6 @@
 package com.aninfo.proyectos.service;
 
+import com.aninfo.proyectos.exception.NoExisteProyectoException;
 import com.aninfo.proyectos.model.Proyecto;
 import com.aninfo.proyectos.model.Recurso;
 import com.aninfo.proyectos.model.Tarea;
@@ -26,8 +27,11 @@ public class ProyectoService {
         }
     }
 
-    public Proyecto getProyecto(int id) {
-        return proyectoRepository.findById(id).get();
+    public Proyecto getProyecto(int id) throws NoExisteProyectoException {
+        if (proyectoRepository.findById(id).isPresent()){
+            return proyectoRepository.findById(id).get();
+        }
+        throw new NoExisteProyectoException();
     }
 
     public ArrayList<Proyecto> getAllProyectos() {
@@ -43,7 +47,10 @@ public class ProyectoService {
         proyectoRepository.deleteById(id);
     }
 
-    public void addTareaToProyecto(int id, Tarea tarea){
+    public void addTareaToProyecto(int id, Tarea tarea) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id).get();
         proyecto.addTarea(tarea);
         tarea.setIdProyecto(id);
@@ -51,7 +58,10 @@ public class ProyectoService {
         proyectoRepository.save(proyecto);
     }
 
-    public void addRecursoToProyecto(int id, Recurso recurso) {
+    public void addRecursoToProyecto(int id, Recurso recurso) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id).get();
         proyecto.addRecurso(recurso);
         recurso.setIdProyecto(id);
@@ -59,21 +69,30 @@ public class ProyectoService {
         proyectoRepository.save(proyecto);
     }
 
-    public void deleteTareaFromProyecto(int id_proyecto, int id_tarea) {
+    public void deleteTareaFromProyecto(int id_proyecto, int id_tarea) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id_proyecto).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id_proyecto).get();
         proyecto.deleteTarea(id_tarea);
         tareaService.deleteTarea(id_tarea);
         proyectoRepository.save(proyecto);
     }
 
-    public void deleteTareaRecursoProyecto(int id_proyecto, int id_recurso) {
+    public void deleteTareaRecursoProyecto(int id_proyecto, int id_recurso) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id_proyecto).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id_proyecto).get();
         proyecto.deleteRecurso(id_recurso);
         recursoService.deleteRecurso(id_recurso);
         proyectoRepository.save(proyecto);
     }
 
-    public void updateTareaFromProyecto(int id_proyecto, int id_tarea, Tarea tarea) {
+    public void updateTareaFromProyecto(int id_proyecto, int id_tarea, Tarea tarea) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id_proyecto).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id_proyecto).get();
         tareaService.updateTarea(id_tarea, tarea);
         proyecto.deleteTarea(id_tarea);
@@ -81,7 +100,10 @@ public class ProyectoService {
         proyectoRepository.save(proyecto);
     }
 
-    public void updateRecursoFromProyecto(int id_proyecto, int id_recurso, Recurso recurso) {
+    public void updateRecursoFromProyecto(int id_proyecto, int id_recurso, Recurso recurso) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id_proyecto).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id_proyecto).get();
         recursoService.updateRecurso(id_recurso, recurso);
         proyecto.deleteRecurso(id_recurso);
@@ -89,12 +111,18 @@ public class ProyectoService {
         proyectoRepository.save(proyecto);
     }
 
-    public ArrayList<Tarea> getAllTareasFromProyecto(int id){
+    public ArrayList<Tarea> getAllTareasFromProyecto(int id) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id).get();
         return proyecto.getTareas();
     }
 
-    public ArrayList<Recurso> getAllRecursosFromProyecto(int id){
+    public ArrayList<Recurso> getAllRecursosFromProyecto(int id) throws NoExisteProyectoException {
+        if (!proyectoRepository.findById(id).isPresent()){
+            throw new NoExisteProyectoException();
+        }
         Proyecto proyecto = proyectoRepository.findById(id).get();
         return proyecto.getRecursos();
     }
